r.cost: fix heap
git-svn-id: https://svn.osgeo.org/grass/grass/branches/releasebranch_7_0@62040 15284696-431f-4ddb-bdfa-cd5b030d7da7

+++ raster/r.cost/cost.h
@@ -37,7 +37,7 @@
 struct cost
 {
     double min_cost;
-    unsigned int age;
+    long age;
     int row;
     int col;
 };
